loadSymbols func added
